requirements.txt
Django==2.2.17
djangorestframework==3.12.4
black==21.8b0
isort==5.9.3
